refactor: add AccountIcon
