+++ packages/shared/src/UI/components/index.ts
@@ -1,3 +1,4 @@
+export * from './AccountIcon/index.js'
 export * from './AddressItem/index.js'
 export * from './AddressViewer/index.js'
 export * from './ApplicationEntry/index.js'
